+++ smallrye-reactive-messaging-kafka/src/test/java/io/smallrye/reactive/messaging/kafka/client/ReactiveKafkaBatchConsumerTest.java
@@ -1,21 +1,9 @@
 package io.smallrye.reactive.messaging.kafka.client;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.Assert.assertTrue;
-import static org.junit.jupiter.api.Assertions.fail;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.Semaphore;
-import java.util.concurrent.TimeUnit;
 
-import org.apache.kafka.clients.consumer.Consumer;
-import org.apache.kafka.common.TopicPartition;
 import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -32,10 +20,6 @@ import io.smallrye.reactive.messaging.test.common.config.MapBasedConfig;
 
 public class ReactiveKafkaBatchConsumerTest extends ClientTestBase {
 
-    private final Semaphore assignSemaphore = new Semaphore(partitions);
-    private final List<Cancellable> subscriptions = new ArrayList<>();
-    private KafkaSource<Integer, String> source;
-
     @AfterEach
     public void tearDown() {
         cancelSubscriptions();
@@ -46,11 +30,6 @@ public class ReactiveKafkaBatchConsumerTest extends ClientTestBase {
     public void init() {
         topic = usage.createNewTopic("test-" + UUID.randomUUID().toString(), partitions);
         resetMessages();
-    }
-
-    private void cancelSubscriptions() {
-        subscriptions.forEach(Cancellable::cancel);
-        subscriptions.clear();
     }
 
     @Test
@@ -64,7 +43,7 @@ public class ReactiveKafkaBatchConsumerTest extends ClientTestBase {
             int receiveStartIndex, int receiveCount, int batchCount) throws Exception {
 
         CountDownLatch latch = new CountDownLatch(batchCount);
-        subscribe(stream, latch);
+        subscribeBatch(stream, latch);
         if (sendCount > 0) {
             sendMessages(sendStartIndex, sendCount);
         }
@@ -90,29 +69,7 @@ public class ReactiveKafkaBatchConsumerTest extends ClientTestBase {
         return source;
     }
 
-    private KafkaConsumerRebalanceListener getKafkaConsumerRebalanceListenerAwaitingAssignation() {
-        return new KafkaConsumerRebalanceListener() {
-            @Override
-            public void onPartitionsAssigned(Consumer<?, ?> consumer,
-                    Collection<TopicPartition> partitions) {
-                ReactiveKafkaBatchConsumerTest.this.onPartitionsAssigned(partitions);
-            }
-        };
-    }
-
-    private void onPartitionsAssigned(Collection<TopicPartition> partitions) {
-        assertThat(topic).isEqualTo(partitions.iterator().next().topic());
-        assignSemaphore.release(partitions.size());
-    }
-
-    private void waitForMessages(CountDownLatch latch) throws InterruptedException {
-        if (!latch.await(receiveTimeoutMillis, TimeUnit.MILLISECONDS)) {
-            fail(latch.getCount() + " messages not received, received=" + count(receivedMessages) + " : "
-                    + receivedMessages);
-        }
-    }
-
-    private void subscribe(Multi<IncomingKafkaRecordBatch<Integer, String>> stream, CountDownLatch... latches)
+    private void subscribeBatch(Multi<IncomingKafkaRecordBatch<Integer, String>> stream, CountDownLatch... latches)
             throws Exception {
         Cancellable cancellable = stream
                 .onItem().invoke(record -> {
@@ -125,12 +82,7 @@ public class ReactiveKafkaBatchConsumerTest extends ClientTestBase {
                     // Ignored.
                 });
         subscriptions.add(cancellable);
-        waitFoPartitionAssignment();
-    }
-
-    private void waitFoPartitionAssignment() throws InterruptedException {
-        Assertions.assertTrue(
-                assignSemaphore.tryAcquire(sessionTimeoutMillis + 1000, TimeUnit.MILLISECONDS), "Partitions not assigned");
+        waitForPartitionAssignment();
     }
 
 }
